revert test

+++ a-v2.0/iserver_revert.sh
@@ -7,8 +7,10 @@ cur_dir=$(cd "$(dirname "$0")"; pwd)
 
 ## »¹Ô­
 function rollbackIServer(){
-  rm -rf $ISERVER_UPLOAD_DIR/*
-  unzip -oq $ISERVER_DEPLOY_BAKUP_DIR/files.zip -d $ISERVER_UPLOAD_DIR
+  if [ -f "$ISERVER_DEPLOY_BAKUP_DIR/files.zip" ]; then
+    rm -rf $ISERVER_UPLOAD_DIR/*
+    unzip -oq $ISERVER_DEPLOY_BAKUP_DIR/files.zip -d $ISERVER_UPLOAD_DIR
+  fi
   rm -rf $TOMCAT_HOME/webapps/ROOT/*
   unzip -oq $ISERVER_DEPLOY_BAKUP_DIR/$ISERVER_PAKAGE_NAME -d $TOMCAT_HOME/webapps/ROOT
   ln -sf $ISERVER_UPLOAD_DIR $TOMCAT_HOME/webapps/ROOT
